haze: swarm!

+++ fleet/modules/ssh-initrd.nix
@@ -4,10 +4,12 @@
     udhcpc.enable = true;
 
     postCommands = ''
-      # Import all pools
-      zpool import -a
-      # Add the load-key command to the .profile
-      echo "zfs load-key -a; killall zfs" >> /root/.profile
+      if type "zpool" > /dev/null; then
+        # Import all pools
+        zpool import -a
+        # Add the load-key command to the .profile
+        echo "zfs load-key -a; killall zfs" >> /root/.profile
+      fi
     '';
 
     ssh = {
